LP-74 Move userDetailsService responsibility to user service
