--- webservice/src/main/java/pl/edu/pjatk/lnpayments/webservice/auth/UserDetailsServiceImpl.java
@@ -1,25 +1,0 @@
-package pl.edu.pjatk.lnpayments.webservice.auth;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import pl.edu.pjatk.lnpayments.webservice.auth.repository.UserRepository;
-
-@Service
-class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Autowired
-    public UserDetailsServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return new UserDetailsImpl(userRepository.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException(username + "not found!")));
-    }
-}
