+++ webservice/src/main/java/pl/edu/pjatk/lnpayments/webservice/auth/config/PasswordEncoderConfig.java
@@ -1,0 +1,16 @@
+package pl.edu.pjatk.lnpayments.webservice.auth.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+class PasswordEncoderConfig {
+
+    @Bean
+    public PasswordEncoder getPasswordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+}
