+++ webservice/src/main/java/pl/edu/pjatk/lnpayments/webservice/auth/service/UserService.java
@@ -1,7 +1,11 @@
 package pl.edu.pjatk.lnpayments.webservice.auth.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import pl.edu.pjatk.lnpayments.webservice.auth.UserDetailsImpl;
 import pl.edu.pjatk.lnpayments.webservice.auth.converter.UserConverter;
 import pl.edu.pjatk.lnpayments.webservice.auth.repository.UserRepository;
 import pl.edu.pjatk.lnpayments.webservice.auth.resource.dto.RegisterRequest;
@@ -11,9 +15,8 @@ import pl.edu.pjatk.lnpayments.webservice.common.entity.User;
 import javax.transaction.Transactional;
 import javax.validation.ValidationException;
 
-
 @Service
-public class UserService {
+public class UserService implements UserDetailsService {
 
     private final UserRepository userRepository;
     private final UserConverter userConverter;
@@ -33,5 +36,11 @@ public class UserService {
 
         User user = userConverter.convertToEntity(request, Role.ROLE_USER);
         userRepository.save(user);
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return new UserDetailsImpl(userRepository.findByEmail(username)
+                .orElseThrow(() -> new UsernameNotFoundException(username + "not found!")));
     }
 }
